refactor(http): split transport/http/desktop.py into a package
The 587-line grab-bag router becomes transport/http/desktop/ with one
module per concern: runtime (shutdown), mcp_clients, ai_provider, sync,
os_open, import_workspace, plus config_store for the shared config.json
helpers. The package __init__ aggregates the routers and re-exports the
existing public surface (router, register_shutdown,
_validated_member_path, _mcp_client_registry) so app.py,
desktop_main.py, and routes/sync.py are untouched.

_backend_stdio_command's parents[2] became parents[3] (one directory
deeper). Tests that monkeypatch _open_with_os/_IS_WINDOWS now import the
owning submodule so the patches keep landing on the namespace the routes
read.
